+++ paramdb/__init__.py
@@ -1,25 +1,19 @@
-"""
-Database for storing and retrieving QPU parameters during quantum control experiments.
-"""
+"""Python package for storing and retrieving experiment parameters."""
 
-from paramdb._keys import CLASS_NAME_KEY, PARAMLIST_ITEMS_KEY, LAST_UPDATED_KEY
 from paramdb._param_data._param_data import ParamData
-from paramdb._param_data._dataclasses import Param, Struct
+from paramdb._param_data._dataclasses import ParamDataclass
 from paramdb._param_data._collections import ParamList, ParamDict
 from paramdb._param_data._type_mixins import ParentType, RootType
-from paramdb._database import ParamDB, CommitEntry, CommitEntryWithData
+from paramdb._database import CLASS_NAME_KEY, ParamDB, CommitEntry, CommitEntryWithData
 
 __all__ = [
-    "CLASS_NAME_KEY",
-    "PARAMLIST_ITEMS_KEY",
-    "LAST_UPDATED_KEY",
     "ParamData",
-    "Param",
-    "Struct",
+    "ParamDataclass",
     "ParamList",
     "ParamDict",
     "ParentType",
     "RootType",
+    "CLASS_NAME_KEY",
     "ParamDB",
     "CommitEntry",
     "CommitEntryWithData",